refactor: move global.ts to theme folder
